Create README.md
Building a logistic regression model and a decision tree based on data from the National Institute of Diabetes and Digestive and Kidney Diseases. The purpose of the dataset is diagnostic to predict whether a patient has diabetes, based on certain diagnostic measurements included in the dataset. Several restrictions were imposed on the selection of these instances from a larger database. In particular, all the patients here are women at least 21 years old of Indian Pima origin.2